Still trying to solve the string conversion problem between wchar_t and char.
git-svn-id: https://irrlichtnetcp.svn.sourceforge.net/svnroot/irrlichtnetcp@80 2f455fab-421c-0410-8db0-c3f11262bd97

+++ IrrlichtW/animatedmesh.cpp
@@ -65,7 +65,7 @@ void AnimatedMeshSceneNode_SetAnimationEndCallback(IntPtr node, ANIMATIONENDCALL
 
 void AnimatedMeshSceneNode_SetAnimationSpeed(IntPtr node, int framePS)
 {
-    GetNodeFromIntPtr(node)->setAnimationSpeed(framePS);
+    GetNodeFromIntPtr(node)->setAnimationSpeed((float)framePS);
 }
 
 void AnimatedMeshSceneNode_SetCurrentFrame(IntPtr node, int cf)
